Autofield for dicoproduce & commande_produit

+++ siteConfinement/ConfinAide/models.py
@@ -21,7 +21,7 @@ class Produit(models.Model):
         return self.name_Pretty
 
 class DicoProduce(models.Model):
-    id_dico = models.BigIntegerField(primary_key = True)
+    id_dico = models.AutoField(primary_key = True)
     id_produce = models.ForeignKey(Produit, on_delete = models.CASCADE)
     quantity = models.IntegerField()
     def __str__(self):
@@ -34,7 +34,7 @@ STATUS_CHOICES = [
 ]
 
 class commande_produit(models.Model):
-    id_commande_produit = models.BigIntegerField(primary_key = True)
+    id_commande_produit = models.AutoField(primary_key = True)
     name_produit = models.ManyToManyField(DicoProduce)
     id_client = models.ForeignKey(Client, on_delete = models.CASCADE)
     confirm = models.CharField(default = "En attente",max_length=30, choices=STATUS_CHOICES)
